Fixed style and changed file names
